move added VM funcs to VM feature
