+++ conoha_client/add_vm/domain/__init__.py
@@ -1,4 +1,3 @@
 """ADD VM feature domain."""
-from .added_vm import AddedVM  # noqa: F401
 from .domain import *  # noqa: F403
 from .errors import *  # noqa: F403
