--- conoha_client/add_vm/domain/added_vm.py
@@ -1,10 +1,0 @@
-"""a Domain Model."""
-from uuid import UUID
-
-from pydantic import BaseModel, Field
-
-
-class AddedVM(BaseModel, frozen=True):
-    """新規追加されたVM."""
-
-    vm_id: UUID = Field(alias="id")
