+++ conoha_client/add_vm/domain/errors.py
@@ -3,11 +3,3 @@
 
 class ImageIdentifyError(Exception):
     """VM Imageを一意に特定できなかった."""
-
-
-class NotFlavorProvidesError(Exception):
-    """そのImageとFlavorの組み合わせは提供されていない."""
-
-
-class NotFoundAddedVMError(Exception):
-    """VMが追加されたはずなのに契約中VM一覧にはない."""
