+++ conoha_client/features/vm/errors.py
@@ -1,0 +1,9 @@
+"""VM errors."""
+
+
+class NotFlavorProvidesError(Exception):
+    """そのImageとFlavorの組み合わせは提供されていない."""
+
+
+class NotFoundAddedVMError(Exception):
+    """VMが追加されたはずなのに契約中VM一覧にはない."""
