+++ conoha_client/features/vm/repo/__init__.py
@@ -1,0 +1,2 @@
+"""VM repo."""
+from .query import get_dep, list_servers  # noqa: F401
