+++ conoha_client/features/vm/repo/query.py
@@ -4,8 +4,7 @@ from __future__ import annotations
 from typing import Callable
 
 from conoha_client.features._shared import Endpoints
-
-from .domain import Server
+from conoha_client.features.vm.domain import VM
 
 
 def get_dep() -> list[object]:
@@ -15,8 +14,8 @@ def get_dep() -> list[object]:
 
 
 def list_servers(
-    get: Callable[[], list[object]] = get_dep,
-) -> list[Server]:
+    dep: Callable[[], list[object]] = get_dep,
+) -> list[VM]:
     """契約中のサーバー情報一覧を取得する."""
-    res = get()
-    return [Server.model_validate(e) for e in res]
+    res = dep()
+    return [VM.model_validate(e) for e in res]
